rewrote base controller and updated base model

+++ api/System/MVC/Controller.php
@@ -9,13 +9,7 @@ namespace MVC;
 use \Exception as Exception;
 use PDOException;
 
-/**
- * Class Controller, a port of MVC
- *
- * @author Mohammad Rahmani <rto1680@gmail.com>
- *
- * @package MVC
- */
+
 class Controller {
 
     /**
@@ -27,7 +21,7 @@ class Controller {
      * Response Class.
      */
     public $response;
-    private $_model;
+    protected $_model;
 	/**
 	*  Construct
 	*/
@@ -35,34 +29,51 @@ class Controller {
         $this->request = $GLOBALS['request'];
         $this->response = $GLOBALS['response'];
         $model = str_replace('Controller', '', get_class($this));
+                
         $this->_model = $this->model($model);
     }
     public function get_all($params=null){
         try{
+            if ($params){
+                foreach($params as $key=>$value){
+                    $this->_model->where($key, $value);
+                }
+            }
             $data = $this->_model->search();
             if (count($data))
                 $this->send(200, ['response'=>'OK', 'data'=>$data]);
             else 
-                $this->send(204, ['response'=> 'No content']);
+                $this->send(404, ['response'=> 'No resource found']);
         }
         catch(PDOException $e){
             $this->send(400, ['response'=> $e->getMessage()]);
         } 
     }
 
-    public function get($id){
-        
-        try{
-            $this->_model->id = $id;
-            $data = $this->_model->search();
-            if (count($data))
-                $this->send(200, ['response'=>'OK', 'data'=>$data]);
-            else 
-                $this->send(204, ['response'=> 'No content']);
+    public function get($params){
+        if (!$params){
+            $this->send(400, ['error'=>'Bad request']);
         }
-        catch(PDOException $e){
-            $this->send(400, ['response'=> $e->getMessage()]);
-        } 
+        else {
+            $id = $params['id'];
+            try{
+                if (count($params) > 1){
+                    foreach($params as $key=>$value){
+                        $this->_model->where($key, $value);
+                    }
+                }
+                $this->_model->id = $id;
+                $data = $this->_model->search();
+                if (count($data))
+                    $this->send(200, ['response'=>'OK', 'data'=>$data]);
+                else 
+                    $this->send(404, ['response'=> 'No resource found']);
+            }
+            catch(PDOException $e){
+                $this->send(400, ['response'=> $e->getMessage()]);
+            } 
+        }
+
         
     }
 
@@ -79,29 +90,43 @@ class Controller {
         } 
 
     }
-    public function update($id){
-        $data = json_decode(file_get_contents('php://input'), true);
-        try{
-            $this->_model->setAtrributes($data);
-            $this->_model->id = $id;
-            $this->_model->save();
-            $this->send(200, ['response'=>'OK']);
+    public function update($params){
+        if (!$params){
+            $this->send(400, ['error' => 'Bad Request']);
         }
-        catch(PDOException $e){
-            $this->send(400, ['response'=> $e->getMessage()]);
-        } 
+        else {
+            $id = $params['id'];
+            $data = json_decode(file_get_contents('php://input'), true);
+            try{
+                $this->_model->setAtrributes($data);
+                $this->_model->id = $id;
+                $this->_model->save();
+                $this->send(200, ['response'=>'OK']);
+            }
+            catch(PDOException $e){
+                $this->send(400, ['response'=> $e->getMessage()]);
+            } 
+        }
+
 
     }
 
-    public function delete($id){
-        try{  
-            $this->_model->id = $id;
-            $this->_model->delete();
-            $this->send(200, ['response'=>'OK']);
+    public function delete($params){
+        if (!$params){
+            $this->send(400, ['error' => 'Bad Request']);
         }
-        catch(PDOException $e){
-            $this->send(400, ['response'=> $e->getMessage()]);
-        } 
+        else {
+            $id = $params['id'];
+            try{  
+                $this->_model->id = $id;
+                $this->_model->delete();
+                $this->send(200, ['response'=>'OK']);
+            }
+            catch(PDOException $e){
+                $this->send(400, ['response'=> $e->getMessage()]);
+            } 
+        }
+
     }
 
     /**
@@ -114,14 +139,14 @@ class Controller {
     private function model($model) {
         // echo $model . PHP_EOL;
         $file = MODELS . ucfirst($model) . '.php';
-        // echo "Inside controller: ".$file . PHP_EOL;
+
 		// check exists file
         if (file_exists($file)) {
             require_once $file;
             
        
             $model = ucfirst($model) . 'Model';
-
+            
 			// check class exists
             if (class_exists($model)){
                 // echo "Inside controller - Model: $model" . PHP_EOL;
